Change folder structure
* Add additional test case for block comments
