+++ src/sloc.js
