+++ test/test_assets/multipleblockstarts.c
@@ -1,0 +1,8 @@
+/* 
+/* <- additional block comment start 
+comment
+*/
+code
+code 
+code
+/* comment */
